Move PluginGroup related traits from plugin to plugin_group
